--- a/src/Hashids.net/Alphabets/AlphabetProvider.cs
+++ b/src/Hashids.net/Alphabets/AlphabetProvider.cs
@@ -19,8 +19,8 @@ namespace HashidsNet.Alphabets
         public virtual IAlphabet GetAlphabet(int index)
         {
             ISalt salt = Salt
-                 .Create(_default.GetChar(index))
-                 .Concat(_salt);
+                .Create(_default.GetChar(index))
+                .Concat(_salt);
 
             return new CharsAlphabet(salt, _chars);
         }