--- a/src/Hashids.net/Alphabets/AlphabetProvider.cs
+++ b/src/Hashids.net/Alphabets/AlphabetProvider.cs
@@ -5,29 +5,31 @@ namespace HashidsNet.Alphabets
 {
     public class AlphabetProvider : IAlphabetProvider
     {
-        private readonly char[] _chars;
         private readonly ISalt _salt;
         private readonly CharsAlphabet _default;
 
         public AlphabetProvider(char[] chars, char[] salt)
         {
-            _chars = chars;
-            _salt = Salt.Create(salt).Snapshot();
-            _default = new CharsAlphabet(_salt, chars);
+            Chars = chars;
+            Salt = Salts.Salt.Create(salt).Snapshot();
+            Default = new CharsAlphabet(Salt, chars);
         }
 
         public virtual IAlphabet GetAlphabet(int index)
         {
-            return LotteryAlphabet.Get(_default, _default.GetChar(index), _salt);
+            return LotteryAlphabet.Get(_default, _default.GetChar(index), Salt);
         }
 
         public IAlphabet GetAlphabet(char @char)
         {
-            int index = Array.IndexOf(_chars, @char);
+            int index = Array.IndexOf(Chars, @char);
 
             return index >= 0 ? GetAlphabet(index) : null;
         }
 
-        public IAlphabet Default => _default;
+        public IAlphabet Default { get; }
+
+        protected char[] Chars { get; }
+        protected ISalt Salt { get; }
     }
 }