--- a/src/Hashids.net/Alphabets/CharsAlphabet.cs
+++ b/src/Hashids.net/Alphabets/CharsAlphabet.cs
@@ -1,0 +1,55 @@
+﻿using HashidsNet.Alphabets.Salts;
+using System;
+
+namespace HashidsNet.Alphabets
+{
+    public sealed class CharsAlphabet : IAlphabet
+    {
+        private readonly ISalt _salt;
+        private readonly char[] _chars;
+
+        public CharsAlphabet(ISalt salt, char[] chars)
+        {
+            _salt = salt;
+            _chars = new char[chars.Length];
+
+            Array.Copy(chars, 0, _chars, 0, chars.Length);
+        }
+
+        public char GetChar(int index)
+        {
+            return _chars[index];
+        }
+
+        public int GetIndex(char @char)
+        {
+            return Array.IndexOf(_chars, @char);
+        }
+
+        public void CopyChars(char[] buffer, int sourceIndex, int destinationIndex, int length)
+        {
+            Array.Copy(_chars, sourceIndex, buffer, destinationIndex, length);
+        }
+
+        public IAlphabet NextPage()
+        {
+            _salt.Concat(_chars).Shuffle(_chars);
+
+            return this;
+        }
+
+        public IAlphabet NextShuffle()
+        {
+            Salt.Create(_chars).Shuffle(_chars);
+
+            return this;
+        }
+
+        public IAlphabet Clone()
+        {
+            return new CharsAlphabet(_salt, _chars);
+        }
+
+        public int Length => _chars.Length;
+    }
+}