--- a/src/Hashids.net/Alphabets/Salts/CharSalt.cs
+++ b/src/Hashids.net/Alphabets/Salts/CharSalt.cs
@@ -4,13 +4,6 @@ namespace HashidsNet.Alphabets.Salts
 {
     public class CharSalt : ISalt
     {
-        private readonly char _char;
-
-        public CharSalt(char @char)
-        {
-            _char = @char;
-        }
-
         public void Calculate(Span<int> buffer, int saltIndex, ref int saltSum)
         {
             if (buffer.Length == 0)
@@ -19,9 +12,11 @@ namespace HashidsNet.Alphabets.Salts
             if (buffer.Length > 1)
                 throw new InvalidOperationException("The one char salt could calculate only sinle value.");
 
-            buffer[0] = saltSum + saltIndex + _char + _char;
-            saltSum += _char;
+            buffer[0] = saltSum + saltIndex + Char + Char;
+            saltSum += Char;
         }
+
+        public char Char { get; set; }
 
         public int Length => 1;
     }