--- a/src/Hashids.net/Alphabets/Salts/CharsSalt.cs
+++ b/src/Hashids.net/Alphabets/Salts/CharsSalt.cs
@@ -5,10 +5,13 @@ namespace HashidsNet.Alphabets.Salts
     public sealed class CharsSalt : ISalt
     {
         private readonly char[] _chars;
+        private int _length;
 
         public CharsSalt(char[] chars)
         {
             _chars = chars;
+
+            Length = _chars.Length;
         }
 
         public void Calculate(Span<int> buffer, int saltIndex, ref int saltSum)
@@ -16,8 +19,8 @@ namespace HashidsNet.Alphabets.Salts
             if (buffer.Length == 0)
                 return;
 
-            if (buffer.Length > _chars.Length)
-                throw new InvalidOperationException($"The salt could calculate only { _chars.Length} values.");
+            if (buffer.Length > _length)
+                throw new InvalidOperationException($"The salt could calculate only {_length} values.");
 
             for (var i = 0; i < buffer.Length; i++)
             {
@@ -29,6 +32,16 @@ namespace HashidsNet.Alphabets.Salts
             }
         }
 
-        public int Length => _chars.Length;
+        public int Length
+        {
+            get { return _length; }
+            set
+            {
+                if (value > _chars.Length)
+                    throw new IndexOutOfRangeException();
+
+                _length = value;
+            }
+        }
     }
 }