--- a/src/Hashids.net/Alphabets/Salts/ISalt.cs
+++ b/src/Hashids.net/Alphabets/Salts/ISalt.cs
@@ -1,0 +1,9 @@
+﻿namespace HashidsNet.Alphabets.Salts
+{
+    public interface ISalt
+    {
+        void Calculate(int[] buffer, int bufferIndex, int saltIndex, int saltLength, ref int saltSum);
+
+        int Length { get; }
+    }
+}